A lot of gubbins involving git and git-svn just occured, largely because
when git-merging these two branches together, git-svn started believing
that these changes were to be commited on top of trunk. Which would have
been bad.

Anyway; this commit pulls in the latest esbmc changes to the doirep form
/ style / branch.

+++ ansi-c/headers/stddef.h
@@ -7,7 +7,7 @@ typedef long int ptrdiff_t;
 
 typedef short wchar_t;
 
-typedef unsigned long int size_t;
+typedef unsigned int size_t;
 
 #define NULL ((void *)0)
